Ajuste de Página Inicial com menu de Produtos | Ajuste de Arquitetura Carrossel -> Produtos | Ajuste em UTF8 das categorias do carrossel
