+++ E-Conc/E-Conc/Controllers/PedidoController.cs
@@ -1,5 +1,4 @@
 ﻿using E_Conc.Data.Interfaces;
-using E_Conc.Enum;
 using E_Conc.Models.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -10,7 +9,6 @@ namespace E_Conc.Controllers
     {
         private readonly IProdutoRepository _produtoRepo;
         private readonly IItemPedidoRespository _itemPedidoRepo;
-        private CarrosselViewModel _carrosselViewModel;
 
         public PedidoController(IProdutoRepository produtoRepo,
                                 IItemPedidoRespository itemPedidoRepo)
@@ -18,35 +16,7 @@ namespace E_Conc.Controllers
             _produtoRepo = produtoRepo;
             _itemPedidoRepo = itemPedidoRepo;
         }
-
-        [Authorize]
-        public IActionResult Carrossel()
-        {
-            var produtosDesenvolvimento = new CategoriaViewModel
-                (_produtoRepo.GetProdutosPorCategoria(Categoria.Desenvolvimento));
-
-            var produtosEmpreendedorismo = new CategoriaViewModel
-                (_produtoRepo.GetProdutosPorCategoria(Categoria.Empreendedorismo));
-
-            var produtosIniciacaoCientifica = new CategoriaViewModel
-                (_produtoRepo.GetProdutosPorCategoria(Categoria.IniciacaoCientifica));
-
-            var produtosPesquisaAcademica = new CategoriaViewModel
-                (_produtoRepo.GetProdutosPorCategoria(Categoria.PesquisaAcademica));
-
-            CategoriaViewModel[] _categoriaViewModel = new CategoriaViewModel[]
-            {
-                produtosDesenvolvimento,
-                produtosEmpreendedorismo,
-                produtosIniciacaoCientifica,
-                produtosPesquisaAcademica
-            };
-            
-            _carrosselViewModel = new CarrosselViewModel(_categoriaViewModel);
-
-            return View(_carrosselViewModel);
-        }
-
+        
         [Authorize]
         public IActionResult Carrinho(int? produtoId)
         {
