+++ E-Conc/E-Conc/Controllers/ProdutoController.cs
@@ -1,4 +1,6 @@
 ﻿using E_Conc.Data.Interfaces;
+using E_Conc.Enum;
+using E_Conc.Models.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -7,10 +9,39 @@ namespace E_Conc.Controllers
     public class ProdutoController : Controller
     {
         private readonly IProdutoRepository _produtoRepo;
+        private CarrosselViewModel _carrosselViewModel;
 
         public ProdutoController(IProdutoRepository produtoRepo)
         {
             _produtoRepo = produtoRepo;
+        }
+
+        [Authorize]
+        public IActionResult Carrossel()
+        {
+            var produtosDesenvolvimento = new CategoriaViewModel
+                (_produtoRepo.GetProdutosPorCategoria(Categoria.Desenvolvimento));
+
+            var produtosEmpreendedorismo = new CategoriaViewModel
+                (_produtoRepo.GetProdutosPorCategoria(Categoria.Empreendedorismo));
+
+            var produtosIniciacaoCientifica = new CategoriaViewModel
+                (_produtoRepo.GetProdutosPorCategoria(Categoria.IniciacaoCientifica));
+
+            var produtosPesquisaAcademica = new CategoriaViewModel
+                (_produtoRepo.GetProdutosPorCategoria(Categoria.PesquisaAcademica));
+
+            CategoriaViewModel[] _categoriaViewModel = new CategoriaViewModel[]
+            {
+                produtosDesenvolvimento,
+                produtosEmpreendedorismo,
+                produtosIniciacaoCientifica,
+                produtosPesquisaAcademica
+            };
+
+            _carrosselViewModel = new CarrosselViewModel(_categoriaViewModel);
+
+            return View(_carrosselViewModel);
         }
 
         [Authorize]
@@ -25,9 +56,6 @@ namespace E_Conc.Controllers
             {
                 return RedirectToAction("Pedido", "Carrossel");
             }
-                
-
-            
         }
 
         [Authorize]
