+++ E-Conc/E-Conc/Models/ViewModels/CarrosselViewModel.cs
@@ -1,12 +1,15 @@
-﻿namespace E_Conc.Models.ViewModels
+﻿using E_Conc.Enum;
+
+namespace E_Conc.Models.ViewModels
 {
     public class CarrosselViewModel
     {
         public CategoriaViewModel[] CategoriasViewModel { get; set; }
+        public Categoria Categoria { get; private set; }
 
         public CarrosselViewModel(CategoriaViewModel[] categoriasViewModel)
         {
             CategoriasViewModel = categoriasViewModel;
-        }       
+        }
     }
 }
